+++ app/src/main/java/com/example/rentit/presentation/home/createpost/components/TagButton.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.home.createpost.component
+package com.example.rentit.presentation.home.createpost.components
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
