+++ app/src/main/java/com/example/rentit/presentation/mypage/MyPageScreen.kt
@@ -58,7 +58,7 @@ import com.example.rentit.common.theme.pretendardFamily
 import com.example.rentit.data.product.dto.ProductDto
 import com.example.rentit.data.user.dto.ReservationDto
 import com.example.rentit.common.component.ProductListItem
-import com.example.rentit.presentation.mypage.component.MyRentalHistoryListItem
+import com.example.rentit.presentation.mypage.components.MyRentalHistoryListItem
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
